5 only one level of sub machine supported (#74)
* modify printAncestry to print multiple generations, with selectors for capitalization and printing self

* adapted tests to corrected hungarianToUnderbarCaps routine; fixed test makefiles so that CFLAGS does not -D

* use -MG to add missing headers to dependency list

* all tests pass with new generateFSMOutputGeneratorGenerator

* Parser uses factories to get output generators. Factories know the difference between top-level and sub-machine generators.  Factories return pointer to static struct for top-level generators.

* -M option added; works for -th and -tp

* multiple fsms on command line working for HTML

* printAncestry now correctly handles including self when self is the top of the tree.  HTML output now prints ancestry in the header.

* plantuml puts ancestry in the requested title

* C machines now do -M; Event enumeration and cross reference now do multiple sub machine levels.

* One sub-level machines now pass all tests; have begun work on second level.

* This does not build; commit only for safe keeping.

* interim; does not build

* another interim

* keep track of the ordinal of each list member

* yet another unbuildable interim

* Added the concept of public and private header files.  That part works, but we're not yet complete.

* interim; for backup

* all tests now pass; there is some cleanup work left, though, so there will be at least one more commit

* removed trailing CRs

* remove vi swap file

* another interim waypoint; replacing direct calls to ancestry and hungarian functions with c language parts functions; all tests pass

* another interim; all tests pass

* interim state function generation; tests pass

* interim print_void_returning_state_fn_case; tests pass

* print_state_returning_state_fn_case; tests pass

* cswitchMachineHeaderEnd; tests pass

* cswitchSubMachineHeaderEnd; tests pass

* writeActionsReturnStateSwitchFSM; tests pass

* new make assisting options work; need to finish adding to help output; tests pass

* unit test for new -M related options

* updated help with new file listing options

* use 'rm -f' to avoid useless error reporting

* add html and plantuml to file list tests; remove for loop from final recipe - failures are hidden

* don't capture result so that it will show up in github's build log

* do not define FDMLANG_DEVELOP; address cmake build warnings

* fix over-zealous correction of cmake build warning

* tweak 'good' handling

* speed up clean target with -j

* accomodate removal of FSMLANG_DEVELOP define

* debugging github build

* remove the multiple target for now

* add comment

* add missing linefeed

* add sub-machine depth tracker

* don't print test execution line

* adjust canonicals for corrected sub-machine depth guage

* add function to get total sub-machine depth

* add option to reduce the user function (e.g. action functions) name prefix to the machine name

* adapted to the new UFMN macro which facilitates short or long user function name prefix

* deprecate the --generate-run-function option, since a run function is always generated and no RUN_STATE_MACHINE macro is provided

* Sigh.  Removed obviated help line.

* use the new UFMN macro for functions

* revert whitespace only changes

* implement review suggestions

* put the new construction in the right place - look for the size of the file after the file has been written

* adjust for FSMLang version 1.41

* external event references no longer used

* finally resolved the troubles with this test by removing the external event reference in this file

* create and use static utility function to create a string from the contents of a file

* use fsm source and header list generation

* clean test before removing fsm

* Correct output redirection

* bump revision in anticipation of release

---------

+++ src/fsm_html.c
@@ -48,9 +48,13 @@
 /*
 	Our interface to the outside world
 */
-int initHTMLWriter(pFSMOutputGenerator,char *);
-void writeHTMLWriter(pFSMOutputGenerator,pMACHINE_INFO);
-void closeHTMLWriter(pFSMOutputGenerator,int);
+static int initHTMLWriter(pFSMOutputGenerator,char *);
+static void writeHTMLWriter(pFSMOutputGenerator,pMACHINE_INFO);
+static void closeHTMLWriter(pFSMOutputGenerator,int);
+
+static int initHTMLFileNameWriter(pFSMOutputGenerator,char *);
+static void writeHTMLFileName(pFSMOutputGenerator,pMACHINE_INFO);
+static void closeHTMLFileNameWriter(pFSMOutputGenerator,int);
 
 typedef struct _fsm_html_output_generator_ FSMHTMLOutputGenerator, *pFSMHTMLOutputGenerator;
 typedef struct _html_machine_data_ HTMLMachineData, *pHTMLMachineData;
@@ -69,7 +73,7 @@ struct _fsm_html_output_generator_
    pHTMLMachineData   pmd;
 };
 
-FSMHTMLOutputGenerator HTMLMachineWriter = {
+static FSMHTMLOutputGenerator HTMLMachineWriter = {
 	{
      initHTMLWriter,
      writeHTMLWriter,
@@ -77,18 +81,6 @@ FSMHTMLOutputGenerator HTMLMachineWriter = {
   },
   NULL
 };
-
-FSMHTMLOutputGenerator HTMLSubMachineWriter = {
-	{
-     initHTMLWriter,
-     writeHTMLWriter,
-     closeHTMLWriter
-  },
-  NULL
-};
-
-pFSMOutputGenerator pHTMLMachineWriter    = (pFSMOutputGenerator) &HTMLMachineWriter;
-pFSMOutputGenerator pHTMLSubMachineWriter = (pFSMOutputGenerator) &HTMLSubMachineWriter;
 
 bool  css_content_internal = false;
 char *css_content_filename = "fsmlang.css";
@@ -100,9 +92,9 @@ static bool print_machine_name_as_list_element(pLIST_ELEMENT pelem, void *data)
    FILE         *fout = (FILE*) data;
 
    fprintf(fout
-           , "<li>%s</li>\n"
-           , pmi->name->name
-           );
+		   , "<li>%s</li>\n"
+		   , pmi->name->name
+		   );
 
    return false;
 }
@@ -195,7 +187,8 @@ static bool print_sub_machine_row(pLIST_ELEMENT pelem, void *data)
            , pmi->name->name
            , pmi->name->name
            , pmi->name->name
-           );
+		   );
+
    fprintf(pfsmhtmlog->pmd->htmlFile
            ,"<td>\n"
            );
@@ -210,7 +203,22 @@ static bool print_sub_machine_row(pLIST_ELEMENT pelem, void *data)
 }
 
 /* Main section */
-int initHTMLWriter (pFSMOutputGenerator pfsmog, char *baseFileName)
+static int initHTMLFileNameWriter (pFSMOutputGenerator pfsmog, char *baseFileName)
+{
+	pFSMHTMLOutputGenerator pfsmhtmlog = (pFSMHTMLOutputGenerator) pfsmog;
+	if (NULL != (pfsmhtmlog->pmd = calloc(1, sizeof(HTMLMachineData))))
+	{
+		if (baseFileName)
+		{
+			pfsmhtmlog->pmd->htmlName = createFileName(baseFileName,".html");
+		}
+	}
+
+	/* this may look funny, but it does the trick */
+	return ((int) !pfsmhtmlog->pmd->htmlName);
+}
+
+static int initHTMLWriter (pFSMOutputGenerator pfsmog, char *baseFileName)
 {
 
   pFSMHTMLOutputGenerator pfsmhtmlog = (pFSMHTMLOutputGenerator) pfsmog;
@@ -280,13 +288,25 @@ int initHTMLWriter (pFSMOutputGenerator pfsmog, char *baseFileName)
 
 }
 
-void writeHTMLWriter(pFSMOutputGenerator pfsmog, pMACHINE_INFO pmi)
+static void writeHTMLFileName(pFSMOutputGenerator pfsmog, pMACHINE_INFO pmi)
+{
+	pFSMHTMLOutputGenerator pfsmhtmlog = (pFSMHTMLOutputGenerator)pfsmog;
+
+	printf("%s ", pfsmhtmlog->pmd->htmlName);
+
+	if (pmi->machine_list)
+	{
+		write_machines(pmi->machine_list, generateHTMLMachineWriter, pfsmog);
+	}
+}
+
+static void writeHTMLWriter(pFSMOutputGenerator pfsmog, pMACHINE_INFO pmi)
 {
 
 	pID_INFO			  pid;
 	unsigned				e,s;
 
-  pFSMHTMLOutputGenerator pfsmhtmlog = (pFSMHTMLOutputGenerator) pfsmog;
+	pFSMHTMLOutputGenerator pfsmhtmlog = (pFSMHTMLOutputGenerator)pfsmog;
 
 	if (!pmi)
 
@@ -294,9 +314,9 @@ void writeHTMLWriter(pFSMOutputGenerator pfsmog, pMACHINE_INFO pmi)
 
   pfsmhtmlog->pmd->pmi = pmi;
 
-	fprintf(pfsmhtmlog->pmd->htmlFile,"<h2>%s</h2>\n"
-		, pmi->name->name
-		);
+	fprintf(pfsmhtmlog->pmd->htmlFile,"<h2>");
+	printAncestry(pmi, pfsmhtmlog->pmd->htmlFile, "::", alc_lower, ai_include_self);
+	fprintf(pfsmhtmlog->pmd->htmlFile, "</h2>\n");
 
 	if (pmi->name->docCmnt)
 		fprintf(pfsmhtmlog->pmd->htmlFile,"<p>%s</p>\n",pmi->name->docCmnt);
@@ -687,29 +707,83 @@ void writeHTMLWriter(pFSMOutputGenerator pfsmog, pMACHINE_INFO pmi)
 
      fprintf(pfsmhtmlog->pmd->htmlFile, "</table>\n");
 
-     write_machines(pmi->machine_list, pHTMLSubMachineWriter);
+	 write_machines(pmi->machine_list, generateHTMLMachineWriter, pfsmog);
   }
 }
 
-void closeHTMLWriter(pFSMOutputGenerator pfsmog, int good)
+static void closeHTMLFileNameWriter(pFSMOutputGenerator pfsmog, int good)
 {
-   pFSMHTMLOutputGenerator pfsmhtmlog = (pFSMHTMLOutputGenerator) pfsmog;
+	pFSMHTMLOutputGenerator pfsmhtmlog = (pFSMHTMLOutputGenerator) pfsmog;
 
-	if (good) {
+	(void) good;
 
-		fprintf(pfsmhtmlog->pmd->htmlFile,"</body>\n</html>\n");
-
-	}
-
-	fclose(pfsmhtmlog->pmd->htmlFile);
-
-	if (!good) {
-
-		unlink(pfsmhtmlog->pmd->htmlName);
-
+	if (!pfsmhtmlog || !pfsmhtmlog->pmd)
+	{
+		return;
 	}
 
 	CHECK_AND_FREE(pfsmhtmlog->pmd->htmlName);
 
+	free(pfsmhtmlog->pmd);
+}
+
+static void closeHTMLWriter(pFSMOutputGenerator pfsmog, int good)
+{
+   pFSMHTMLOutputGenerator pfsmhtmlog = (pFSMHTMLOutputGenerator) pfsmog;
+
+   if (!pfsmhtmlog || !pfsmhtmlog->pmd)
+   {
+	   return;
+   }
+
+   if (good)
+   {
+
+	   fprintf(pfsmhtmlog->pmd->htmlFile, "</body>\n</html>\n");
+
+   }
+
+   fclose(pfsmhtmlog->pmd->htmlFile);
+
+   if (!good)
+   {
+
+	   unlink(pfsmhtmlog->pmd->htmlName);
+
+   }
+
+   CHECK_AND_FREE(pfsmhtmlog->pmd->htmlName);
+
+   free(pfsmhtmlog->pmd);
+
+}
+
+pFSMOutputGenerator generateHTMLMachineWriter(pFSMOutputGenerator parent)
+{
+	pFSMOutputGenerator pfsmog;
+
+	if (parent)
+	{
+		pFSMHTMLOutputGenerator pfsmhtmlog = calloc(1, sizeof(FSMHTMLOutputGenerator));
+
+		pfsmhtmlog->fsmog.writeMachine = writeHTMLWriter;
+		pfsmhtmlog->fsmog.initOutput   = initHTMLWriter;
+		pfsmhtmlog->fsmog.closeOutput  = closeHTMLWriter;
+
+		pfsmog = (pFSMOutputGenerator)pfsmhtmlog;
+	}
+	else
+	{
+		pfsmog = (pFSMOutputGenerator)&HTMLMachineWriter;
+	}
+
+	if (output_generated_file_names_only)
+	{
+		pfsmog->writeMachine = writeHTMLFileName;
+		pfsmog->initOutput   = initHTMLFileNameWriter;
+		pfsmog->closeOutput  = closeHTMLFileNameWriter;
+	}
+
+	return pfsmog;
 }
